+++ docs/changelog.rst
@@ -8,7 +8,7 @@ Changes in 0.5
 
 Changes in 0.4
 ==============
-- Added CRSF support and validate_on_save via flask.ext.WTF
+- Added CSRF support and validate_on_save via flask.ext.WTF
 - Fixed DateTimeField not required
 
 Changes in 0.3