+++ docs/changelog.rst
@@ -2,9 +2,14 @@
 Changelog
 =========
 
+Changes in 0.5
+==============
+- Added support for all connection settings
+- Fixed extended DynamicDocument
+
 Changes in 0.4
 ==============
-- Added CRSF support and validate_on_save via flask.ext.WTF
+- Added CSRF support and validate_on_save via flask.ext.WTF
 - Fixed DateTimeField not required
 
 Changes in 0.3